move altsearch into a different file
